+++ src/org/ods/component/RolloutOpenShiftDeploymentStage.groovy
@@ -82,7 +82,7 @@ class RolloutOpenShiftDeploymentStage extends Stage {
             config.tailorSelector = config.selector
         }
         if (!config.containsKey('tailorVerify')) {
-            config.tailorVerify = false
+            config.tailorVerify = true
         }
         if (!config.containsKey('tailorExclude')) {
             config.tailorExclude = 'bc,is'